+++ LiveFeedBatch/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java-library'
 }
 
-version = '0.1.19'
+version = '0.1.20'
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-batch'
