+++ LiveFeedBatch/src/main/java/com/livefeed/livefeedbatch/batch/common/driver/FirefoxDriverProvider.java
@@ -1,0 +1,34 @@
+package com.livefeed.livefeedbatch.batch.common.driver;
+
+import lombok.extern.slf4j.Slf4j;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.firefox.FirefoxOptions;
+
+@Slf4j
+public class FirefoxDriverProvider {
+
+    private static final String driverPath = getDriverPath();
+    private static final FirefoxOptions firefoxOptions = setFirefoxOptions();
+
+    public static WebDriver getDriver() {
+        System.setProperty("webdriver.gecko.driver", driverPath);
+        System.setProperty("webdriver.firefox.port", "16000");
+        return new FirefoxDriver(firefoxOptions);
+    }
+
+    private static String getDriverPath() {
+        // mac / docker container(linux) 환경 구분
+        String osName = System.getProperty("os.name").toLowerCase();
+        if (osName.contains("mac")) {
+            return "/opt/homebrew/bin/geckodriver";
+        } else {
+            return "/usr/bin/geckodriver";
+        }
+    }
+
+    private static FirefoxOptions setFirefoxOptions() {
+        return new FirefoxOptions()
+                .addArguments("--headless");
+    }
+}
